requirements.txt
fastapi
uvicorn
python-multipart
gtts
sentence-transformers
OpenAI
faiss-cpu
np
SpeechRecognition
pydantic
dotenv
hf_xet
google-cloud-storage
